+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Deploy default and support custom affinities ([#391]).
+- Cluster status conditions ([#403])
 - Extend cluster resources for status and cluster operation (paused, stopped) ([#405])
 
 ### Changed
@@ -31,6 +32,7 @@ All notable changes to this project will be documented in this file.
 [#380]: https://github.com/stackabletech/trino-operator/pull/380
 [#391]: https://github.com/stackabletech/trino-operator/pull/391
 [#400]: https://github.com/stackabletech/trino-operator/pull/400
+[#403]: https://github.com/stackabletech/trino-operator/pull/403
 [#405]: https://github.com/stackabletech/trino-operator/pull/405
 [#406]: https://github.com/stackabletech/trino-operator/pull/406